Define BaseMaterializedWork in normal model code so we don't have to do weird things to import it.
